pyproject.toml
[project]
name = "studyclaw"
version = "0.1.0"
description = "AI-powered learning assistant with RAG and GitHub sync"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",

    # LangChain & LangGraph
    "langchain>=0.1.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=2.0.0",
    "langchain-community>=0.0.10",
    "langchain-chroma>=1.1.0",

    # LLM & Embeddings
    "dashscope>=1.14.0",

    # Vector store
    "chromadb>=0.4.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",

    # Utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.15.4",
    "pypdf>=6.7.5",
    "pydantic>=2.0.0",
    "requests>=2.31.0",

    # GitHub integration
    "PyGithub>=2.0.0",

    # Session persistence
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[dependency-groups]
dev = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
